refactor(redis): replace fxa-shared/redis with vanilla ioredis
because:

More transactional operations are coming with the move of oauth tokens
into redis that the existing `update` function is not well equipped to
handle. By moving transactional operations into lua we can also simplify
our connection handling because we'll no longer require a pool for serial
message isolation. Converting the existing uses of `update` to lua helps
prove out the concept.

Moving the custom serialization of the session tokens into lua allows
the server logic to be agnostic about the storage layer details.