feat(cwts): Add CWTS to the /signup page in Sync optional flow
Add CWTS on the /signup page as an experiment, 20% total,
10% in each bucket.

If user deselects all of the buckets, then does not send
Sync information to the browser in the login message.

As part of this, did a little refactor to how WebChannel
events are stored for testing, instead of always storing
them, store events only if in an automated testing
environment. This made me feel more comfortable storing
the event metadata into sessionStorage too, where it can
be queried from within the functional tests.

fixes #3490